Start over from a scratch
The idea here is that the first implementations was a result of my
starting with mobile developpement with expo and I have made a lot
of stuff here and there so clean up would have a pain. So I decived
to just start over and see if I improved my level.